refactor: decompose TransactionConfirmationModal (#2571)
* refactor: move CurrencyLogo to common/pure

* refactor: move CurrencyLogo hooks

* refactor: move CurrencySelectButton

* refactor: move CurrencyAmountPreview

* refactor: clean up FiatValue props

* refactor: extract PriceImpactIndicator component

* refactor: move FiatValue and PriceImpactIndicator to common pure

* refactor: remove CurrencyInputPanelMod.tsx

* refactor: align LimitOrdersConfirm imports

* refactor: rename areCurrenciesLoading

* chore: fix import

* refactor: move LimitOrdersConfirm to modules/trade

* refactor: rename LimitOrdersConfirm to TradeConfirmation

* refactor: extract limit-orders specific code from TradeConfirmation

* refactor: narrow down CurrencyAmountPreview props

* chore: pull develop

* chore: fix cosmos build (circular dependency)

* chore: cosmos for LimitOrdersConfirmModal

* chore: fix ConfirmationPendingContent cosmos

* chore: cosmos for OrderProgressBar

* chore: extract TransactionSubmittedContentProps component

* chore: component OrderSubmittedContent

* chore: component TradeConfirmModal

* chore: refactor TransactionErrorContent

* refactor: remove L2Content

* refactor: decompose TransactionConfirmationModal
