+++ src/common/pure/AnimatedConfirmation/index.tsx
