+++ src/legacy/components/TransactionConfirmationModal/index.cosmos.tsx
@@ -2,10 +2,12 @@ import { OrderClass, OrderKind, SupportedChainId } from '@cowprotocol/cow-sdk'
 
 import styled from 'styled-components/macro'
 
-import { ConfirmationPendingContent, OperationType } from 'legacy/components/TransactionConfirmationModal'
+import { ConfirmOperationType } from 'legacy/components/TransactionConfirmationModal'
 import { COW, GNO } from 'legacy/constants/tokens'
 import store from 'legacy/state'
 import { addPendingOrder, OrderStatus } from 'legacy/state/orders/actions'
+
+import { LegacyConfirmationPendingContent } from './LegacyConfirmationPendingContent'
 
 const txHash = '0xe87e1d02b052daa9605abe018e8172feffd1bc38ed2284e6hhhhhh' + Date.now()
 
@@ -17,7 +19,7 @@ const defaultProps = {
   hash: txHash,
   pendingText: <div>pendingText</div>,
   currencyToAdd: COW[SupportedChainId.MAINNET],
-  operationType: OperationType.ORDER_SIGN,
+  operationType: ConfirmOperationType.ORDER_SIGN,
   inline: false,
 }
 
@@ -33,7 +35,7 @@ const Wrapper = styled.div`
 const Fixtures = {
   ConfirmationPendingContent: (
     <Wrapper>
-      <ConfirmationPendingContent {...defaultProps} />
+      <LegacyConfirmationPendingContent {...defaultProps} />
     </Wrapper>
   ),
 }
