+++ src/legacy/components/TransactionConfirmationModal/types.ts
@@ -1,0 +1,10 @@
+export enum ConfirmOperationType {
+  WRAP_ETHER,
+  UNWRAP_WETH,
+  APPROVE_TOKEN,
+  REVOKE_APPROVE_TOKEN,
+  ORDER_SIGN,
+  ORDER_CANCEL,
+  CONVERT_VCOW,
+  CLAIM_VESTED_COW,
+}
