+++ src/legacy/hooks/useTransactionConfirmationModal.tsx
@@ -1,20 +1,19 @@
 import { useState, useCallback } from 'react'
 
-import TransactionConfirmationModal from 'legacy/components/TransactionConfirmationModal'
-import { OperationType } from 'legacy/components/TransactionConfirmationModal'
+import { TransactionConfirmationModal, ConfirmOperationType } from 'legacy/components/TransactionConfirmationModal'
 import { useOpenModal, useCloseModals, useModalIsOpen } from 'legacy/state/application/hooks'
 import { ApplicationModal } from 'legacy/state/application/reducer'
 
 export default function useTransactionConfirmationModal(
-  defaultOperationType: OperationType = OperationType.WRAP_ETHER
+  defaultOperationType: ConfirmOperationType = ConfirmOperationType.WRAP_ETHER
 ) {
-  const [operationType, setOperationType] = useState<OperationType>(defaultOperationType)
+  const [operationType, setOperationType] = useState<ConfirmOperationType>(defaultOperationType)
   const [transactionConfirmationModalMsg, setTransactionConfirmationModalMsg] = useState<string>()
   const openTransactionConfirmationModalAux = useOpenModal(ApplicationModal.TRANSACTION_CONFIRMATION)
   const closeModal = useCloseModals()
   const showTransactionConfirmationModal = useModalIsOpen(ApplicationModal.TRANSACTION_CONFIRMATION)
   const openModal = useCallback(
-    (message: string, operationType: OperationType) => {
+    (message: string, operationType: ConfirmOperationType) => {
       setTransactionConfirmationModalMsg(message)
       setOperationType(operationType)
       openTransactionConfirmationModalAux()
