+++ src/legacy/hooks/useWrapCallback.ts
@@ -6,7 +6,8 @@ import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 import { t } from '@lingui/macro'
 
 import { wrapAnalytics } from 'legacy/components/analytics'
-import { getOperationMessage, OperationType } from 'legacy/components/TransactionConfirmationModal'
+import { ConfirmOperationType } from 'legacy/components/TransactionConfirmationModal'
+import { getOperationMessage } from 'legacy/components/TransactionConfirmationModal/LegacyConfirmationPendingContent'
 import { RADIX_HEX } from 'legacy/constants'
 import { useWETHContract } from 'legacy/hooks/useContract'
 import { useCloseModals } from 'legacy/state/application/hooks'
@@ -42,7 +43,7 @@ export type WrapUnwrapCallback = (params?: WrapUnwrapCallbackParams) => Promise<
 
 type TransactionAdder = ReturnType<typeof useTransactionAdder>
 
-export type OpenSwapConfirmModalCallback = (message: string, operationType: OperationType) => void
+export type OpenSwapConfirmModalCallback = (message: string, operationType: ConfirmOperationType) => void
 
 export interface WrapUnwrapContext {
   wrapType: WrapType
@@ -114,11 +115,11 @@ export function useWrapUnwrapContext(
     return null
   }
 
-  const openTransactionConfirmationModal = (pendingText: string, operationType: OperationType) => {
+  const openTransactionConfirmationModal = (pendingText: string, operationType: ConfirmOperationType) => {
     openTxConfirmationModal({ operationType, pendingText })
   }
   const amountHex = `0x${inputAmount.quotient.toString(RADIX_HEX)}`
-  const operationType = isWrap ? OperationType.WRAP_ETHER : OperationType.UNWRAP_WETH
+  const operationType = isWrap ? ConfirmOperationType.WRAP_ETHER : ConfirmOperationType.UNWRAP_WETH
   const baseSummarySuffix = isWrap ? `${native} to ${wrapped}` : `${wrapped} to ${native}`
   const baseSummary = `${formatTokenAmount(inputAmount)} ${baseSummarySuffix}`
   const summary = `${isWrap ? 'Wrap' : 'Unwrap'} ${baseSummary}`
@@ -170,7 +171,7 @@ export async function wrapUnwrapCallback(
     closeModals,
   } = context
   const isWrap = wrapType === WrapType.WRAP
-  const operationType = isWrap ? OperationType.WRAP_ETHER : OperationType.UNWRAP_WETH
+  const operationType = isWrap ? ConfirmOperationType.WRAP_ETHER : ConfirmOperationType.UNWRAP_WETH
 
   try {
     useModals && openTransactionConfirmationModal(confirmationMessage, operationType)
